Add files via upload
Thrilled to receive this certificate for my participation in Spectra 3.0 by the Codebusters Club at GLA University! 🎉 This experience was a fantastic opportunity to dive deeper into computer engineering and applications, connect with like-minded peers, and grow my skills.

Grateful to everyone who made this event possible. Excited to keep pushing forward in my journey in technology and engineering! 🚀

#Spectra3 #Codebusters #GLAUniversity #Achievement #EngineeringJourney #ElectronicsEngineering #KeepLearning